chore: rename utils to common
