+++ backend-ping-pong/src/common/Ball.ts
