+++ backend-ping-pong/src/common/IMessageBrocker.ts
