+++ backend-ping-pong/src/common/Player.ts
