+++ backend-ping-pong/src/common/Rectangle.ts
