+++ backend-ping-pong/src/duel/MessageBrocker.ts
@@ -1,5 +1,5 @@
 import { FastifyInstance } from "fastify";
-import { IMessageBroker } from "../utils/IMessageBrocker";
+import { IMessageBroker } from "../common/IMessageBrocker";
 
 export class MessageBrocker implements IMessageBroker {
     constructor(private fastify: FastifyInstance) { }
