+++ backend-ping-pong/src/routes/duel.ts
@@ -1,8 +1,8 @@
 import { FastifyPluginAsync } from "fastify"
 import websocket from '@fastify/websocket'
 import { v4 as uuidv4 } from 'uuid'
-import { MatchManager } from "../utils/MatchManager"
-import { Player } from "../utils/Player"
+import { MatchManager } from "../common/MatchManager"
+import { Player } from "../common/Player"
 
 const example: FastifyPluginAsync = async (fastify, opts): Promise<void> => {
     await fastify.register(websocket)
